refactor: separar factory DI de subscrição de eventos (C6)
Extrai subscricao-eventos-rodada.ts do rodada-factory.ts, separando
duas responsabilidades que estavam misturadas:

- subscricao-eventos-rodada.ts: CallbacksRodada + subscreverEventos
- rodada-factory.ts: DI pura (criarDecisores, fabricarRodada, fabricarPartida)

Nenhuma mudança em consumidores (JogoScene). Factory reexporta
CallbacksRodada para manter compatibilidade.
